+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * Fix initialization of new instances to avoid overwriting needed value and causing later error.
 * Upgrade detection of AltUI so we don't falsely detect when bridged (on "real" device triggers AltUI feature registration).
-* Make use of Reactor, if installed, to detect house mode changes more quickly (this does not require additional ReactorSensors or configuration on the user's part, it just starts working when DEMII and Reactor are installed together). The HouseModes plugin also works, but is less responsive than Reactor.
+* Use Reactor's HMT approach for super-responsive house mode tracking (Reactor does not need to be installed).
 
 ## Version 2.8 (released) ##